Add typed requests
