+++ projects/vehicle/src/lib/components/create-update-maintenance-item-drawer/create-update-maintenance-item-lateral-drawer.component.ts
@@ -102,7 +102,7 @@ export class CreateUpdateMaintenanceItemLateralDrawerComponent
       this.maintenanceItem == null
         ? this.vehicleService.postCreateMaintenanceItemAsync(request)
         : this.vehicleService.putUpdateMaintenanceItemAsync(
-            this.maintenanceItem!.id,
+            this.maintenanceItem.id,
             request,
           );
