+++ projects/vehicle/src/lib/models/maintenance-item-create.model.ts
@@ -1,6 +1,3 @@
-export interface MaintenancePlanCreate {
-  vehicleId: number;
-  maintenanceItemId: number;
-  kmInterval: number | null;
-  timeInterval: number | null;
+export interface CreateMaintenanceItemRequest {
+  description: string;
 }
