+++ projects/vehicle/src/lib/models/maintenance-item-update.model.ts
@@ -1,0 +1,3 @@
+export interface UpdateMaintenanceItemRequest {
+  description: string;
+}
