+++ projects/vehicle/src/lib/models/maintenance-plan-create.model.ts
@@ -1,0 +1,6 @@
+export interface MaintenancePlanCreate {
+  vehicleId: number;
+  maintenanceItemId: number;
+  kmInterval: number | null;
+  timeInterval: number | null;
+}
